Part of WOR-165: step 1 — create watcher_types.py, add Import Linter contracts
Move constants, LinearClientProtocol, ActiveWorker, is_watcher_running, and
_to_metrics_mode to watcher_types.py (leaf module). Add both Import Linter
contracts (watcher-layers, watcher-types-is-leaf) to .importlinter. Create
empty stubs for the 4 remaining sub-modules so the layers contract is satisfied
from this first commit onward. Backward-compat re-exports for LinearClientProtocol
and is_watcher_running kept in watcher.py.
