+++ app/core/watcher.py
@@ -27,47 +27,35 @@ import subprocess  # nosec B404
 import sys
 import threading
 import time
-from dataclasses import dataclass, field
 from pathlib import Path
-from typing import IO, Any, Protocol
+from typing import IO
 
 from app.core.escalation_policy import EscalationPolicy
 from app.core.linear_client import DONE_STATE_TYPES, LinearError
 from app.core.manifest import ExecutionManifest
-from app.core.metrics import ImplementationMode, MetricsStore, Outcome, TicketMetrics
-
-logger = logging.getLogger(__name__)
-
-_CLAUDE_DIR = ".claude"
-_PID_FILE = Path(_CLAUDE_DIR) / "watcher.pid"
-_LITELLM_PORT = 8082
-_LITELLM_CONFIG = "litellm-local.yaml"
-_LOCAL_MODEL = "qwen3-coder:30b"
-_LITELLM_BASE_URL = f"http://localhost:{_LITELLM_PORT}"
-_OLLAMA_PORT = 11434
-_OLLAMA_KEEPALIVE = "120m"
-_WORKTREE_BASE = Path("worktrees")
-
-_ENV_VARS_TO_STRIP_FOR_CLOUD = frozenset(
-    {
-        "ANTHROPIC_BASE_URL",
-        "ANTHROPIC_MODEL",
-        "OPENAI_API_BASE",
-    }
+from app.core.metrics import MetricsStore, Outcome, TicketMetrics
+from app.core.watcher_types import (
+    _CLAUDE_DIR,
+    _ENV_VARS_TO_STRIP_FOR_CLOUD,
+    _LITELLM_BASE_URL,
+    _LITELLM_CONFIG,
+    _LITELLM_PORT,
+    _LOCAL_MODEL,
+    _OLLAMA_KEEPALIVE,
+    _OLLAMA_PORT,
+    _PID_FILE,
+    _WORKTREE_BASE,
+    ActiveWorker,
+    _to_metrics_mode,
+)
+from app.core.watcher_types import (
+    LinearClientProtocol as LinearClientProtocol,  # noqa: F401 — backward-compat re-export
+)
+from app.core.watcher_types import (
+    is_watcher_running as is_watcher_running,  # noqa: F401 — backward-compat re-export
 )
 
-
-# ---------------------------------------------------------------------------
-# Protocol for dependency injection (testability)
-# ---------------------------------------------------------------------------
-
-
-class LinearClientProtocol(Protocol):
-    def list_ready_for_local(self) -> list[dict[str, Any]]: ...
-    def get_open_blockers(self, issue_id: str) -> list[str]: ...
-    def set_state(self, issue_id: str, state_name: str) -> None: ...
-    def post_comment(self, issue_id: str, body: str) -> None: ...
-    def get_issue_state_type(self, identifier: str) -> str | None: ...
+logger = logging.getLogger(__name__)
 
 
 # ---------------------------------------------------------------------------
@@ -97,18 +85,6 @@ def _parse_worker_usage(log_path: Path) -> tuple[int | None, int | None]:
     except Exception:
         return None, None
     return None, None
-
-
-@dataclass
-class ActiveWorker:
-    ticket_id: str
-    linear_id: str
-    manifest: ExecutionManifest
-    worktree_path: Path
-    process: subprocess.Popen[bytes]
-    start_time: float = field(default_factory=time.monotonic)
-    backed_up_plans: list[Path] = field(default_factory=list)
-    retry_count: int = 0
 
 
 # ---------------------------------------------------------------------------
@@ -1448,39 +1424,3 @@ class Watcher:
             _PID_FILE.unlink()
         except FileNotFoundError:
             pass
-
-
-# ---------------------------------------------------------------------------
-# Utilities
-# ---------------------------------------------------------------------------
-
-
-def is_watcher_running(pid_file: Path = _PID_FILE) -> bool:
-    """Return True if a watcher process is currently running."""
-    if not pid_file.exists():
-        return False
-    try:
-        pid = int(pid_file.read_text(encoding="utf-8").strip())
-    except (ValueError, OSError):
-        return False
-    # Check if process is alive (cross-platform)
-    if sys.platform == "win32":
-        import ctypes
-
-        handle = ctypes.windll.kernel32.OpenProcess(0x00100000, False, pid)
-        if not handle:
-            return False
-        ctypes.windll.kernel32.CloseHandle(handle)
-        return True
-    else:
-        try:
-            os.kill(pid, 0)
-            return True
-        except (ProcessLookupError, PermissionError):
-            return False
-
-
-def _to_metrics_mode(mode: str) -> ImplementationMode:
-    if mode in ("local", "cloud", "hybrid"):
-        return mode  # type: ignore[return-value]
-    return "cloud"
