+++ app/core/watcher_helpers.py
@@ -1,0 +1,4 @@
+"""Pure helper functions for the watcher sub-system (no I/O, unit-testable).
+
+Populated in WOR-165 step 2.
+"""
