+++ app/core/watcher_services.py
@@ -1,0 +1,4 @@
+"""LiteLLM and Ollama process management for the watcher sub-system.
+
+Populated in WOR-165 step 3.
+"""
