+++ app/core/watcher_subprocess.py
@@ -1,0 +1,4 @@
+"""Subprocess and I/O functions for the watcher sub-system.
+
+Populated in WOR-165 step 5.
+"""
