+++ app/core/watcher_worktrees.py
@@ -1,0 +1,4 @@
+"""Worktree lifecycle functions for the watcher sub-system.
+
+Populated in WOR-165 step 4.
+"""
